+++ Content.Server/Antag/Selectors/LinearAntagCount.cs
@@ -9,6 +9,6 @@ public sealed partial class LinearAntagCount : MinMaxAntagCountSelector
 {
     public override int GetTargetAntagCount(IRobustRandom random, int playerCount)
     {
-        return Math.Clamp(playerCount / PlayerRatio, Range.Min, Range.Max);
+        return (int)Math.Clamp(playerCount / PlayerRatio, Range.Min, Range.Max); // Starlight, float
     }
 }
